Moved creation of text exercise to a new page
