+++ src/main/webapp/app/entities/text-exercise/index.ts
@@ -2,7 +2,6 @@ export * from './text-exercise.model';
 export * from './text-exercise-update.component';
 export * from './text-exercise-popup.service';
 export * from './text-exercise.service';
-export * from './text-exercise-dialog.component';
 export * from './text-exercise-delete-dialog.component';
 export * from './text-exercise-detail.component';
 export * from './text-exercise.component';
