+++ src/main/webapp/app/entities/text-exercise/text-exercise-update.scss
