+++ src/main/webapp/app/entities/text-exercise/text-exercise.module.ts
@@ -9,8 +9,6 @@ import {
     TextExerciseDeleteDialogComponent,
     TextExerciseDeletePopupComponent,
     TextExerciseDetailComponent,
-    TextExerciseDialogComponent,
-    TextExercisePopupComponent,
     textExercisePopupRoute,
     TextExercisePopupService,
     textExerciseRoute,
@@ -35,23 +33,8 @@ const ENTITY_STATES = [...textExerciseRoute, ...textExercisePopupRoute];
         ArtemisDifficultyPickerModule,
         ArtemisMarkdownEditorModule,
     ],
-    declarations: [
-        TextExerciseComponent,
-        TextExerciseDetailComponent,
-        TextExerciseUpdateComponent,
-        TextExerciseDialogComponent,
-        TextExerciseDeleteDialogComponent,
-        TextExercisePopupComponent,
-        TextExerciseDeletePopupComponent,
-    ],
-    entryComponents: [
-        TextExerciseComponent,
-        TextExerciseDialogComponent,
-        TextExerciseUpdateComponent,
-        TextExercisePopupComponent,
-        TextExerciseDeleteDialogComponent,
-        TextExerciseDeletePopupComponent,
-    ],
+    declarations: [TextExerciseComponent, TextExerciseDetailComponent, TextExerciseUpdateComponent, TextExerciseDeleteDialogComponent, TextExerciseDeletePopupComponent],
+    entryComponents: [TextExerciseComponent, TextExerciseDeleteDialogComponent, TextExerciseDeletePopupComponent],
     providers: [TextExerciseService, TextExercisePopupService, { provide: JhiLanguageService, useClass: JhiLanguageService }],
     exports: [TextExerciseComponent],
 })
