+++ src/main/webapp/app/entities/text-exercise/text-exercise.route.ts
@@ -3,7 +3,7 @@ import { Routes } from '@angular/router';
 import { UserRouteAccessService } from '../../core';
 import { TextExerciseComponent } from './text-exercise.component';
 import { TextExerciseDetailComponent } from './text-exercise-detail.component';
-import { TextExercisePopupComponent } from './text-exercise-dialog.component';
+import { TextExerciseUpdateComponent } from './text-exercise-update.component';
 import { TextExerciseDeletePopupComponent } from './text-exercise-delete-dialog.component';
 
 export const textExerciseRoute: Routes = [
@@ -34,29 +34,27 @@ export const textExerciseRoute: Routes = [
         },
         canActivate: [UserRouteAccessService],
     },
+    {
+        path: 'course/:courseId/text-exercise/new',
+        component: TextExerciseUpdateComponent,
+        data: {
+            authorities: ['ROLE_TA', 'ROLE_INSTRUCTOR', 'ROLE_ADMIN'],
+            pageTitle: 'artemisApp.textExercise.home.title',
+        },
+        canActivate: [UserRouteAccessService],
+    },
+    {
+        path: 'text-exercise/:exerciseId/edit',
+        component: TextExerciseUpdateComponent,
+        data: {
+            authorities: ['ROLE_TA', 'ROLE_INSTRUCTOR', 'ROLE_ADMIN'],
+            pageTitle: 'artemisApp.textExercise.home.title',
+        },
+        canActivate: [UserRouteAccessService],
+    },
 ];
 
 export const textExercisePopupRoute: Routes = [
-    {
-        path: 'course/:courseId/text-exercise-new',
-        component: TextExercisePopupComponent,
-        data: {
-            authorities: ['ROLE_TA', 'ROLE_INSTRUCTOR', 'ROLE_ADMIN'],
-            pageTitle: 'artemisApp.textExercise.home.title',
-        },
-        canActivate: [UserRouteAccessService],
-        outlet: 'popup',
-    },
-    {
-        path: 'text-exercise/:id/edit',
-        component: TextExercisePopupComponent,
-        data: {
-            authorities: ['ROLE_TA', 'ROLE_INSTRUCTOR', 'ROLE_ADMIN'],
-            pageTitle: 'artemisApp.textExercise.home.title',
-        },
-        canActivate: [UserRouteAccessService],
-        outlet: 'popup',
-    },
     {
         path: 'text-exercise/:id/delete',
         component: TextExerciseDeletePopupComponent,
